backend/api/filters.py
import django_filters
from recipes.models import Recipe, Tag


class RecipeFilter(django_filters.FilterSet):
    """Фильтр для модели Recipe."""

    tags = django_filters.ModelMultipleChoiceFilter(
        field_name='tags__slug',
        to_field_name='slug',
        queryset=Tag.objects.all()
    )

    class Meta:
        model = Recipe
        fields = ('tags',)
