backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.4
oauthlib==3.2.2
Pillow==9.0.0
pycparser==2.21
PyJWT==2.8.0
pypdf==3.15.5
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zipp==3.16.2
psycopg2-binary==2.9.3
